requirements.txt
mysql-connector-python~=8.0.28
Flask~=2.0.2
opencv-python~=4.5.5.62

uwsgi
psycopg2